boolean instrumentation tag
