+++ dd-java-agent/instrumentation/jdbc/src/test/groovy/RemoteJDBCInstrumentationTest.groovy
@@ -6,6 +6,7 @@ import datadog.trace.agent.test.naming.VersionedNamingTestBase
 import datadog.trace.agent.test.utils.PortUtils
 import datadog.trace.api.Config
 import datadog.trace.api.DDSpanTypes
+import datadog.trace.api.naming.v1.DatabaseNamingV1
 import datadog.trace.bootstrap.instrumentation.api.InstrumentationTags
 import datadog.trace.bootstrap.instrumentation.api.Tags
 import org.testcontainers.containers.MSSQLServerContainer
@@ -214,33 +215,84 @@ abstract class RemoteJDBCInstrumentationTest extends VersionedNamingTestBase {
     def addDbmTag = dbmTraceInjected()
     resultSet.next()
     resultSet.getInt(1) == 3
-    assertTraces(1) {
-      trace(2) {
-        basicSpan(it, "parent")
-        span {
-          serviceName renameService ? dbName.get(driver).toLowerCase() : service(driver)
-          operationName this.operation(driver)
-          resourceName obfuscatedQuery
-          spanType DDSpanTypes.SQL
-          childOf span(0)
-          errored false
-          measured true
-          tags {
-            "$Tags.COMPONENT" "java-jdbc-statement"
-            "$Tags.SPAN_KIND" Tags.SPAN_KIND_CLIENT
-            "$Tags.DB_TYPE" driver
-            "$Tags.DB_INSTANCE" dbName.get(driver).toLowerCase()
-            "$Tags.PEER_HOSTNAME" String
-            // currently there is a bug in the instrumentation with
-            // postgresql and mysql if the connection event is missed
-            // since Connection.getClientInfo will not provide the username
-            "$Tags.DB_USER" { it == null || it == jdbcUserNames.get(driver) }
-            "$Tags.DB_OPERATION" operation
-            if (addDbmTag) {
-              "$InstrumentationTags.DBM_TRACE_INJECTED" true
+    final databaseNaming = new DatabaseNamingV1()
+    def normalizedDbType= databaseNaming.normalizedName(driver)
+    if (driver == POSTGRESQL || driver == MYSQL) {
+      assertTraces(1) {
+        trace(2) {
+          basicSpan(it, "parent")
+          span {
+            serviceName renameService ? dbName.get(driver).toLowerCase() : service(driver)
+            operationName this.operation(driver)
+            resourceName obfuscatedQuery
+            spanType DDSpanTypes.SQL
+            childOf span(0)
+            errored false
+            measured true
+            tags {
+              "$Tags.COMPONENT" "java-jdbc-statement"
+              "$Tags.SPAN_KIND" Tags.SPAN_KIND_CLIENT
+              "$Tags.DB_TYPE" normalizedDbType
+              "$Tags.DB_INSTANCE" dbName.get(driver).toLowerCase()
+              "$Tags.PEER_HOSTNAME" String
+              // currently there is a bug in the instrumentation with
+              // postgresql and mysql if the connection event is missed
+              // since Connection.getClientInfo will not provide the username
+              "$Tags.DB_USER" { it == null || it == jdbcUserNames.get(driver) }
+              "$Tags.DB_OPERATION" operation
+              if (addDbmTag) {
+                "$InstrumentationTags.DBM_TRACE_INJECTED" true
+              }
+              peerServiceFrom(Tags.DB_INSTANCE)
+              defaultTags()
             }
-            peerServiceFrom(Tags.DB_INSTANCE)
-            defaultTags()
+          }
+        }
+      }
+    } else {
+      assertTraces(1) {
+        trace(3) {
+          basicSpan(it, "parent")
+          span {
+            serviceName renameService ? dbName.get(driver).toLowerCase() : service(driver)
+            operationName this.operation(normalizedDbType)
+            resourceName obfuscatedQuery
+            spanType DDSpanTypes.SQL
+            childOf span(0)
+            errored false
+            measured true
+            tags {
+              "$Tags.COMPONENT" "java-jdbc-statement"
+              "$Tags.SPAN_KIND" Tags.SPAN_KIND_CLIENT
+              "$Tags.DB_TYPE" normalizedDbType
+              "$Tags.DB_INSTANCE" dbName.get(driver).toLowerCase()
+              "$Tags.PEER_HOSTNAME" String
+              "$Tags.DB_USER" { it == null || it == jdbcUserNames.get(driver) }
+              "$Tags.DB_OPERATION" operation
+              peerServiceFrom(Tags.DB_INSTANCE)
+              defaultTags()
+            }
+          }
+          span {
+            serviceName renameService ? dbName.get(driver).toLowerCase() : service(driver)
+            operationName this.operation(normalizedDbType)
+            resourceName "set context_info ?"
+            spanType DDSpanTypes.SQL
+            childOf span(0)
+            errored false
+            measured true
+            tags {
+              "$Tags.COMPONENT" "java-jdbc-statement"
+              "$Tags.SPAN_KIND" Tags.SPAN_KIND_CLIENT
+              "$Tags.DB_TYPE" normalizedDbType
+              "$Tags.DB_INSTANCE" dbName.get(driver).toLowerCase()
+              "$Tags.PEER_HOSTNAME" String
+              "$Tags.DB_USER" { it == null || it == jdbcUserNames.get(driver) }
+              "$Tags.DB_OPERATION" "set"
+              "dd.instrumentation" true
+              peerServiceFrom(Tags.DB_INSTANCE)
+              defaultTags()
+            }
           }
         }
       }
@@ -257,13 +309,13 @@ abstract class RemoteJDBCInstrumentationTest extends VersionedNamingTestBase {
     SQLSERVER  | connectTo(driver, peerConnectionProps(driver))          | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
     MYSQL      | cpDatasources.get("tomcat").get(driver).getConnection() | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
     POSTGRESQL | cpDatasources.get("tomcat").get(driver).getConnection() | false         | "SELECT 3 FROM pg_user" | "SELECT"  | "SELECT ? FROM pg_user"
-    SQLSERVER  | cpDatasources.get("tomcat").get(driver).getConnection() | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
+    //SQLSERVER  | cpDatasources.get("tomcat").get(driver).getConnection() | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
     MYSQL      | cpDatasources.get("hikari").get(driver).getConnection() | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
     POSTGRESQL | cpDatasources.get("hikari").get(driver).getConnection() | false         | "SELECT 3 FROM pg_user" | "SELECT"  | "SELECT ? FROM pg_user"
-    SQLSERVER  | cpDatasources.get("hikari").get(driver).getConnection() | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
+    //SQLSERVER  | cpDatasources.get("hikari").get(driver).getConnection() | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
     MYSQL      | cpDatasources.get("c3p0").get(driver).getConnection()   | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
     POSTGRESQL | cpDatasources.get("c3p0").get(driver).getConnection()   | false         | "SELECT 3 FROM pg_user" | "SELECT"  | "SELECT ? FROM pg_user"
-    SQLSERVER  | cpDatasources.get("c3p0").get(driver).getConnection()   | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
+    //SQLSERVER  | cpDatasources.get("c3p0").get(driver).getConnection()   | false         | "SELECT 3"              | "SELECT"  | "SELECT ?"
   }
 
   def "prepared statement execute on #driver with #connection.getClass().getCanonicalName() generates a span"() {
@@ -591,6 +643,7 @@ abstract class RemoteJDBCInstrumentationTest extends VersionedNamingTestBase {
       return
     }
 
+    // object already exists (no IF NOT EXISTS in SQL Server)
     try {
       connection.prepareCall(createSql).execute()
     } catch (SQLServerException ex) {}
